+++ src/withComponents/Room/Room.tsx
@@ -27,6 +27,7 @@ import { useParticipantMetaContext } from '../ParticipantMetaProvider/usePartici
 import { LocationTuple } from '../../types';
 import { useLocalVolumeDetection } from '../../withHooks/useLocalVolumeDetection/useLocalVolumeDetection';
 import { motion } from 'framer-motion';
+import { StickyNoteWidget } from '../StickyNoteWidget/StickyNoteWidget';
 
 export interface DragItem {
   type: string;
@@ -55,9 +56,67 @@ export const Room: React.FC<IRoomProps> = ({ initialAvatar }) => {
 
   useLocalVolumeDetection();
 
-  const widgetLinks = widgets.filter(widget => widget.type === WidgetTypes.Link);
-  // safe to assume only one whiteboard ever
-  const [widgetWhiteboard] = widgets.filter(widget => widget.type === WidgetTypes.Whiteboard);
+  // get the widgets to display in the room
+  const widgetComps = widgets.reduce<JSX.Element[]>((widComps, widget) => {
+    if (widget.data.isPublished || widget.participantSid === localParticipant.sid) {
+      switch (widget.type) {
+        case WidgetTypes.Link:
+          widComps.push(
+            <LinkWidget
+              key={widget.id}
+              id={widget.id}
+              isPublished={widget.data.isPublished}
+              position={widget.location}
+              title={widget.data.title}
+              url={widget.data.url}
+              onCloseHandler={() => removeWidget(widget.id)}
+              participant={
+                widget.participantSid === localParticipant.sid
+                  ? localParticipant
+                  : remoteParticipants.find(pt => pt.sid === widget.participantSid)
+              }
+              dragConstraints={dragableArea}
+              initialOffset={widget.data.initialOffset}
+            />
+          );
+          break;
+        case WidgetTypes.StickyNote:
+          widComps.push(
+            <StickyNoteWidget
+              key={widget.id}
+              id={widget.id}
+              position={widget.location}
+              text={widget.data.text}
+              isPublished={widget.data.isPublished}
+              participant={
+                widget.participantSid === localParticipant.sid
+                  ? localParticipant
+                  : remoteParticipants.find(pt => pt.sid === widget.participantSid)
+              }
+              onCloseHandler={() => removeWidget(widget.id)}
+              dragConstraints={dragableArea}
+              initialOffset={widget.data.initialOffset}
+            />
+          );
+          break;
+        case WidgetTypes.Whiteboard:
+          widComps.push(
+            <Whiteboard
+              widgetId={widget.id}
+              onCloseHandler={() => removeWidget(widget.id)}
+              whiteboardId={widget.data.whiteboardId}
+              dragConstraints={dragableArea}
+              position={widget.location}
+              initialOffset={widget.data.initialOffset}
+            />
+          );
+          break;
+        default:
+          break;
+      }
+    }
+    return widComps;
+  }, []);
 
   useEffect(() => {
     if (initialAvatar) {
@@ -171,36 +230,7 @@ export const Room: React.FC<IRoomProps> = ({ initialAvatar }) => {
         {Object.keys(huddles).map(huddleId => (
           <HuddleBubble huddleId={huddleId} participants={huddles[huddleId]} key={huddleId} />
         ))}
-        <div>
-          {widgetWhiteboard ? (
-            <Whiteboard
-              widgetId={widgetWhiteboard.id}
-              onCloseHandler={() => removeWidget(widgetWhiteboard.id)}
-              whiteboardId={widgetWhiteboard.data.whiteboardId}
-              dragConstraints={dragableArea}
-              position={widgetWhiteboard.location}
-            />
-          ) : null}
-        </div>
-        <div key="widgies" className="u-flex u-flexWrap u-floatRight u-flexJustifyEnd" style={{ maxWidth: '40%' }}>
-          {widgetLinks.map(widget => (
-            <LinkWidget
-              key={widget.id}
-              id={widget.id}
-              position={widget.location}
-              title={widget.data.title}
-              url={widget.data.url}
-              onCloseHandler={() => removeWidget(widget.id)}
-              participant={
-                widget.participantSid === localParticipant.sid
-                  ? localParticipant
-                  : remoteParticipants.find(pt => pt.sid === widget.participantSid)
-              }
-              dragConstraints={dragableArea}
-              classNames={widget.location ? 'u-positionAbsolute' : 'u-positionRelative'}
-            />
-          ))}
-        </div>
+        <div className="widgets">{widgetComps}</div>
         {Object.keys(bubs).map(key => {
           const { pt, top, left } = bubs[key];