+++ src/withHooks/useHandleDeviceChange/useHandleDeviceChange.ts
@@ -4,98 +4,77 @@
  * audio/video track publications.
  */
 
-import { useCallback, useEffect } from 'react';
+import { useEffect } from 'react';
+import Cookie from 'js-cookie';
+
 import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
 import { useParticipantMetaContext } from '../../withComponents/ParticipantMetaProvider/useParticipantMetaContext';
 import { useAVSourcesContext } from '../../withComponents/AVSourcesProvider/useAVSourcesContext';
 
+import { LocalAudioTrack, LocalVideoTrack } from 'twilio-video';
+
 export function useHandleDeviceChange(activeCameraId: string, activeMicId: string) {
   const {
     room: { localParticipant },
-    getLocalVideoTrack,
-    getLocalAudioTrack,
+    localTracks,
   } = useVideoContext();
 
   const devices = useAVSourcesContext();
 
   const { updateActiveCamera, updateActiveMic } = useParticipantMetaContext();
 
-  // Callback to unpublish/publish video tracks to achieve a change in the video source device.
-  const changeActiveCamera = useCallback(
-    (deviceId: string) => {
-      if (localParticipant) {
-        const pubs = Array.from(localParticipant.videoTracks.values());
-
-        pubs.forEach(pub => {
-          localParticipant.unpublishTrack(pub.track);
-          // This event will cause the publications array in the `usePublications` hook to update, thus causing the
-          // Publications rendered in ParticipantCircle to update, thus removing and detaching the video element
-          // previously attached with the other camera.
-          localParticipant.emit('trackUnpublished', pub);
-          pub.track.stop();
-        });
-
-        getLocalVideoTrack(deviceId).then(track => {
-          localParticipant.publishTrack(track);
-        });
-      }
-    },
-    [getLocalVideoTrack, localParticipant]
-  );
+  const audioTrack = localTracks.find(track => track.kind === 'audio') as LocalAudioTrack;
+  const videoTrack = localTracks.find(track => track.kind === 'video') as LocalVideoTrack;
 
   // Effect to unpublish/publish video tracks when the specified active camera id changes.
   useEffect(() => {
-    // Have to base the presence of video on the videoTracks property of the localParticipant, rather than the
-    // `localTracks` property of the videoContext. For some reason while swapping camera video tracks,
-    // videoContext.localTracks doesn't contain any video tracks. Seems that localParticipant.videoTracks may be a
-    // better source of truth? This could be a happy accident that makes camera swapping work. More research may be
-    // necessary to decide if this is the correct approach.
-    if (localParticipant && localParticipant.videoTracks.size) {
-      changeActiveCamera(activeCameraId);
+    if (videoTrack) {
+      videoTrack.restart({ deviceId: activeCameraId });
     }
-  }, [activeCameraId, localParticipant, changeActiveCamera]);
-
-  // Callback to unpublish/publish audio tracks to achieve a change in the audio source device.
-  const changeActiveMic = useCallback(
-    (deviceId: string) => {
-      if (localParticipant) {
-        const pubs = Array.from(localParticipant.audioTracks.values());
-
-        pubs.forEach(pub => {
-          localParticipant.unpublishTrack(pub.track);
-          // This event will cause the publications array in the `usePublications` hook to update, thus causing the
-          // Publications rendered in ParticipantCircle to update, thus removing and detaching the audio element
-          // previously attached with the other mic.
-          localParticipant.emit('trackUnpublished', pub);
-          pub.track.stop();
-        });
-
-        getLocalAudioTrack(deviceId).then(track => {
-          localParticipant.publishTrack(track);
-        });
-      }
-    },
-    [localParticipant, getLocalAudioTrack]
-  );
+  }, [activeCameraId, videoTrack]);
 
   // Effect to unpublish/publish audio tracks when the specified active mic id changes.
   useEffect(() => {
-    if (localParticipant && localParticipant.audioTracks.size) {
-      changeActiveMic(activeMicId);
+    if (audioTrack) {
+      audioTrack.restart({ deviceId: activeMicId });
     }
-  }, [activeMicId, changeActiveMic, localParticipant]);
+  }, [activeMicId, audioTrack]);
 
   // Effect to update the active camera when the available cameras change.
   useEffect(() => {
-    if (localParticipant && !devices.cameras.find(cam => cam.deviceId === activeCameraId)) {
-      updateActiveCamera(devices.cameras[0]?.deviceId || '');
+    if (localParticipant) {
+      // If there is a participant attempt to find a previously set camera id.
+      const prefCamId = Cookie.get('vidPref');
+      const prefCam = prefCamId && devices.cameras.find(cam => cam.deviceId === prefCamId);
+      if (prefCam) {
+        // Previous camera id was set and device is connected. Use that one.
+        updateActiveCamera(prefCam.deviceId);
+      } else {
+        // Otherwise, just grab the first in the list of cameras
+        updateActiveCamera(devices.cameras[0]?.deviceId || '');
+      }
+    } else if (!localParticipant && videoTrack) {
+      // Otherwise if there is video but no participant, just restart the video track so that a camera is working.
+      videoTrack.restart({ deviceId: devices.cameras[0]?.deviceId || '' });
     }
-  }, [localParticipant, devices.cameras, activeCameraId, updateActiveCamera]);
+  }, [localParticipant, devices.cameras, activeCameraId, updateActiveCamera, videoTrack]);
 
   // Effect to update the active mic when the available mics change.
   useEffect(() => {
-    if (localParticipant && !devices.mics.find(mic => mic.deviceId === activeMicId)) {
-      updateActiveMic(devices.mics[0]?.deviceId || '');
+    if (localParticipant) {
+      // If there is a participant attempt to find a previously set mic id.
+      const prefMicId = Cookie.get('micPref');
+      const prefMic = prefMicId && devices.mics.find(mic => mic.deviceId === prefMicId);
+      if (prefMic) {
+        // Previous mic id was set and device is connected. Use that one.
+        updateActiveMic(prefMic.deviceId);
+      } else {
+        // Otherwise, just grab the first in the list of mics.
+        updateActiveMic(devices.cameras[0]?.deviceId || '');
+      }
+    } else if (audioTrack) {
+      // Otherwise, if there is audio but no participant, just restart the audio track so that a mic is working.
+      audioTrack.restart({ deviceId: devices.mics[0]?.deviceId || '' });
     }
-  }, [localParticipant, devices.mics, activeMicId, updateActiveMic]);
+  }, [localParticipant, devices.mics, activeMicId, updateActiveMic, audioTrack]);
 }