+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-## 1.89.0 
+## 1.92.0
+
+- Add in commonsans font
+
+## 1.91.0
+
+- Add in sticky note widget
+- QOL changes for whiteboard and link widgets
+- Widgets now spawn in the middle of the screen
+- Code clean up
+
+## 1.90.0
+
+- Upgrade twilio-video and update local track state management to be …consistent with the latest starter app.
+- Add cookies for mic and camera preferences so that when devices change we can attempt to connect to the last set mic/camera.
+- Only set the mic/camera pref cookie when the user explicitly sets a mic/camera via the settings modal.
+- TODO: fix local audio volume detection. currently stops working if the the audio input device is changed.
+
+## 1.89.0
 
 - Adding in motive room